extensions/telegram_ai_bot/pyproject.toml
[project]
name = "trendradar-telegram-bot"
version = "0.1.0"
description = "Telegram AI bot for Web3 denoise & summarization"
requires-python = ">=3.10"

dependencies = [
  "aiogram>=3.0",
  "feedparser>=6.0",
  "python-dotenv>=1.0",
  "apscheduler>=3.10",
  "aiosqlite>=0.19",
  "aiohttp>=3.9",
  "aiohttp-socks>=0.10.1",
  "litellm>=1.0",
  "cryptography>=41.0"
]
